Refactor Image/ImageGroup stuff

+++ grails-app/domain/be/cytomine/command/abstractimagegroup/DeleteAbstractImageGroupCommand.groovy
@@ -10,36 +10,27 @@ import java.util.prefs.BackingStoreException
 import be.cytomine.image.AbstractImage
 import be.cytomine.security.Group
 import be.cytomine.image.AbstractImageGroup
+import be.cytomine.Exception.CytomineException
+import be.cytomine.Exception.ObjectNotFoundException
 
 class DeleteAbstractImageGroupCommand extends DeleteCommand implements UndoRedoCommand {
 
   boolean saveOnUndoRedoStack = true;
 
-  def execute() {
-    log.info "Execute"
+  def execute() throws CytomineException{
 
-    try {
-      def postData = JSON.parse(postData)
-      AbstractImage abstractimage = AbstractImage.get(postData.abstractimage)
-      Group group = Group.get(postData.group)
+      AbstractImage abstractimage = AbstractImage.get(json.abstractimage)
+      Group group = Group.get(json.group)
 
       log.info "Delete abstractimage-group with abstractimage=" + abstractimage + " group=" + group
 
       AbstractImageGroup abstractimageGroup = AbstractImageGroup.findByAbstractimageAndGroup(abstractimage,group)
-      String id = abstractimageGroup.id
+      if(!abstractimageGroup) throw new ObjectNotFoundException("AbstractImageGroup not found " + abstractimage + " group=" + group)
 
       def response = super.createDeleteMessage(id,abstractimageGroup,[id,abstractimage.id,group.name] as Object[])
       AbstractImageGroup.unlink(abstractimageGroup.abstractimage, abstractimageGroup.group)
 
       return response
-
-    } catch (NullPointerException e) {
-      log.error(e)
-      return [data: [success: false, errors: e.getMessage()], status: 404]
-    } catch (BackingStoreException e) {
-      log.error(e)
-      return [data: [success: false, errors: e.getMessage()], status: 400]
-    }
   }
